LayoutTests/fast/text/variations/out-of-bounds-selection-properties.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<div id="test1" style="font-weight: 1001; font-style: oblique 91deg;">Hello</div>
<div id="test2" style="font-weight: -1; font-stretch: -1%; font-style: oblique -91deg;">Hello</div>
<div id="test3" style="font: 1001 16px 'Times';">Hello</div>
<div id="test4" style="font: -1 16px 'Times';">Hello</div>
<div id="test5" style="font: oblique 91deg 16px 'Times';">Hello</div>
<div id="test6" style="font: oblique -91deg 16px 'Times';">Hello</div>
<script>
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test1')).getPropertyValue('font-weight')", "400");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test1')).getPropertyValue('font-style')", "normal");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test2')).getPropertyValue('font-weight')", "400");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test2')).getPropertyValue('font-stretch')", "100%");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test2')).getPropertyValue('font-style')", "normal");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test3')).getPropertyValue('font-weight')", "400");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test4')).getPropertyValue('font-weight')", "400");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test5')).getPropertyValue('font-style')", "normal");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test6')).getPropertyValue('font-style')", "normal");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
