LayoutTests/fast/text/variations/out-of-bounds-selection-properties-expected.txt

PASS window.getComputedStyle(document.getElementById('test1')).getPropertyValue('font-weight') is "400"
PASS window.getComputedStyle(document.getElementById('test1')).getPropertyValue('font-style') is "normal"
PASS window.getComputedStyle(document.getElementById('test2')).getPropertyValue('font-weight') is "400"
PASS window.getComputedStyle(document.getElementById('test2')).getPropertyValue('font-stretch') is "100%"
PASS window.getComputedStyle(document.getElementById('test2')).getPropertyValue('font-style') is "normal"
PASS window.getComputedStyle(document.getElementById('test3')).getPropertyValue('font-weight') is "400"
PASS window.getComputedStyle(document.getElementById('test4')).getPropertyValue('font-weight') is "400"
PASS window.getComputedStyle(document.getElementById('test5')).getPropertyValue('font-style') is "normal"
PASS window.getComputedStyle(document.getElementById('test6')).getPropertyValue('font-style') is "normal"
PASS successfullyParsed is true

TEST COMPLETE
Hello
Hello
Hello
Hello
Hello
Hello
